requirements.txt
google-generativeai
flask
pymupdf
pytesseract
pdf2image